Extract `PagedRequest` into `TB.DanceDance.API.Contracts.Models` and migrate related DTOs to shared contracts (DD-36).
