+++ src/backend/Application/Features/Comments/Endpoints/ListCommentsByLinkEndpoint.cs
@@ -1,15 +1,10 @@
 using Application.Extensions;
-using Application.Pagination;
 using FastEndpoints;
 using Microsoft.Extensions.Logging;
 using TB.DanceDance.API.Contracts.Features.Comments;
 using TB.DanceDance.API.Contracts.Models;
 
 namespace Application.Features.Comments.Endpoints;
-
-public class ListCommentsByLinkRequest : PagedRequest
-{
-}
 
 /// <summary>
 /// Gets comments for a video accessed through a shared link. Anonymous access allowed.
