+++ src/backend/TB.DanceDance.API.Contracts/Features/Videos/ListMyVideosRequest.cs
@@ -1,0 +1,8 @@
+﻿using TB.DanceDance.API.Contracts.Models;
+
+namespace TB.DanceDance.API.Contracts.Features.Videos
+{
+    public class ListMyVideosRequest : PagedRequest
+    {
+    }
+}
